Avec un hack sale pour le moment (au final ne marchera pas mais pour l'exemple est ok) ca marche pour la bleu, la rouge pas encore mais j'y travaille

+++ src/position.cpp
@@ -2,7 +2,7 @@
 
 namespace rt
 {
-    Position::Position(int _x, int _y, int _z)
+    Position::Position(double _x, double _y, double _z)
     {
         x = _x;
         y = _y;
